Split Leaf/Branch Nodes into separate classes.
Branch nodes need to be much larger than leaf nodes beacuse they contain
polygons. Right now, we pessimistically allocate enough space for branch
nodes even for leaf nodes, because of we use an std::variant to switch
between their types. By breaking them apart, we can reduce memory
consumption by nearly 10x per leaf node, and there are a lot more leaf
nodes than branch nodes.

Everything is hugely broken, but I'm pushing this so I don't lose
these changes in case my computer blows up.

+++ src/storage/tree_node_allocator.cpp
@@ -9,11 +9,11 @@ tree_node_allocator::tree_node_allocator( size_t memory_budget,
         std::string backing_file ) :
     buffer_pool_( memory_budget, backing_file ),
     space_left_in_cur_page_( PAGE_DATA_SIZE ),
-    cur_page_( std::numeric_limits<size_t>::max() ) {
+    cur_page_( std::numeric_limits<uint32_t>::max() ) {
 }
 
-page *tree_node_allocator::get_page_to_alloc_on( size_t object_size ) {
-    if( cur_page_ == std::numeric_limits<size_t>::max() ) {
+page *tree_node_allocator::get_page_to_alloc_on( uint16_t object_size ) {
+    if( cur_page_ == std::numeric_limits<uint32_t>::max() ) {
         cur_page_ = 0;
         page *page_ptr = buffer_pool_.get_page( 0 );
         assert( object_size <= space_left_in_cur_page_ );
